motion: Fix comparison between signed and unsigned interger
If wait_us < 0, comparison against motion_min_interval actually fails,
and this negative wait_us causes task_wait_event() never returns if we
are not using any motion task event except the timer. The motion task
will then stop running and sensor data stay unchanged.

BRANCH=none
BUG=chrome-os-partner:54092
TEST=hardcode wait_us to a negative value before motion_min_interval check,
and see motion task is still running by EC console cmd timerinfo

+++ common/motion_sense.c
@@ -831,7 +831,7 @@ void motion_sense_task(void)
 			 * Guarantee some minimum delay to allow other lower
 			 * priority tasks to run.
 			 */
-			if (wait_us < motion_min_interval)
+			if (wait_us < (int)motion_min_interval)
 				wait_us = motion_min_interval;
 		} else {
 			wait_us = -1;
